Remove the code that enables the DirectX debug layer
Before this commit, starting the app from within Visual Studio and
attaching the debugger resulted in an exception when closing the app.
That is the case because the app terminates without cleaning up the
DirectX resources. This in terms means that the DirectX debug layer
reports memory leaks. This does not seem to be relevant because the app
terminates anyway which is why this commit removes the code that enables
the DirectX debug layer.